lots of stuff in the core (model stuff, sequence clip target value pushing works), some UI work, and a bunch of other things

+++ trunk/Animator/src/anicore/Core/Model/Clips/ClipPropertyData.cs
@@ -56,7 +56,7 @@ namespace Animator.Core.Model.Clips
 			this._Name = (string)element.Attribute(Schema.common_name);
 		}
 
-		public abstract object GetValue(float position);
+		public abstract object GetValue(double position);
 
 		protected IEnumerable<XAttribute> WriteCommonXAttributes()
 		{
